+++ src/lambdacd_git/core.clj
@@ -10,10 +10,16 @@
             [clojure.walk :as walk]
             [ring.middleware.params :as ring-params]
             [ring.util.response :as ring-response]
-            [compojure.core :as compojure])
+            [compojure.core :as compojure]
+            [lambdacd-git.ssh-agent-support :as ssh-agent-support]
+            [me.raynes.fs :as fs])
   (:import (java.util.regex Pattern)
            (java.util Date)
-           (java.text SimpleDateFormat)))
+           (java.text SimpleDateFormat)
+           (org.eclipse.jgit.transport SshSessionFactory)
+           (java.io SequenceInputStream FileInputStream File)
+           (com.jcraft.jsch JSch)
+           (clojure.lang SeqEnumeration)))
 
 (defn- find-changed-revision [old-revisions new-revisions]
   (let [[_ new-entries _] (diff old-revisions new-revisions)
@@ -178,3 +184,24 @@
 
 (defn notifications-for [pipeline]
   (compojure/POST "/notify-git" request (notify-git-handler (:context pipeline) request)))
+
+(defn- known-hosts-streams [known-hosts-files]
+  (->> known-hosts-files
+       (map fs/expand-home)
+       (filter fs/exists?)
+       (map io/file)
+       (map (fn [^File f] (FileInputStream. f)))
+       (SeqEnumeration.)))
+
+(defn- set-known-hosts [known-hosts-files]
+  (fn [^JSch jsch]
+    (doto jsch
+      (.setKnownHosts (SequenceInputStream. (known-hosts-streams known-hosts-files))))))
+
+(defn init-ssh! [& {:keys [use-agent known-hosts-files]
+                    :or   {use-agent         true
+                           known-hosts-files ["~/.ssh/known_hosts" "/etc/ssh/ssh_known_hosts"]}}]
+  (let [customizer-fns (if use-agent
+                         [ssh-agent-support/ssh-agent-customizer (set-known-hosts known-hosts-files)]
+                         [(set-known-hosts known-hosts-files)])]
+    (SshSessionFactory/setInstance (ssh-agent-support/session-factory customizer-fns))))