+++ src/lambdacd_git/ssh_agent_support.clj
@@ -8,25 +8,31 @@
            (com.jcraft.jsch JSch)))
 
 (defn add-ssh-agent-connector [^JSch jsch]
-  (let [usf (JNAUSocketFactory.)
-        con (SSHAgentConnector. usf)
+  (let [usf   (JNAUSocketFactory.)
+        con   (SSHAgentConnector. usf)
         irepo (RemoteIdentityRepository. con)]
     (.setIdentityRepository jsch irepo)))
 
-(defn session-factory []
+(defn ssh-agent-customizer [jsch]
+  (try
+    (if (SSHAgentConnector/isConnectorAvailable)
+      (add-ssh-agent-connector jsch)
+      (log/info "No SSH-Agent connector available. SSH-Keys with passphrases will not be supported"))
+    (catch Exception e
+      (log/warn e "Problems with SSH Agent. Falling back to default behavior"))))
+
+
+; INTERNAL: WILL BE MOVED TO ssh-init
+(defn session-factory [customizer-fns]
   (proxy [JschConfigSessionFactory] []
     (configure [host session]
       ; just to implement the interface
       )
     (createDefaultJSch [^FS fs]
       (let [jsch (proxy-super createDefaultJSch fs)]
-        (try
-          (if (SSHAgentConnector/isConnectorAvailable)
-            (add-ssh-agent-connector jsch)
-            (log/info "No SSH-Agent connector available. SSH-Keys with passphrases will not be supported"))
-          (catch Exception e
-            (log/warn e "Problems with SSH Agent. Falling back to default behavior")))
+        (doall (map #(% jsch) customizer-fns))
         jsch))))
 
+; DEPRECATED: USE lambdacd-git.ssh-init/init-ssh! instead!
 (defn initialize-ssh-agent-support! []
-  (SshSessionFactory/setInstance (session-factory)))
+  (SshSessionFactory/setInstance (session-factory [ssh-agent-customizer])))